+++ apps/preact/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @utrecht/preact-app
 
+## 1.0.41
+
+### Patch Changes
+
+- Updated dependencies [341a0b7]
+  - @utrecht/design-tokens@5.0.1
+
+## 1.0.40
+
+### Patch Changes
+
+- Updated dependencies [d1f8749]
+  - @utrecht/component-library-react@12.0.0
+
+## 1.0.39
+
+### Patch Changes
+
+- Updated dependencies [edac79f]
+- Updated dependencies [b912501]
+- Updated dependencies [0fcdfbf]
+  - @utrecht/component-library-react@11.0.0
+  - @utrecht/design-tokens@5.0.0
+
+## 1.0.38
+
+### Patch Changes
+
+- Updated dependencies [f02b554]
+- Updated dependencies [55988ac]
+  - @utrecht/design-tokens@4.2.0
+  - @utrecht/component-library-react@10.3.10
+
+## 1.0.37
+
+### Patch Changes
+
+- Updated dependencies [e7b59b3]
+- Updated dependencies [951bb2c]
+- Updated dependencies [fd2db7b]
+  - @utrecht/design-tokens@4.1.1
+
 ## 1.0.36
 
 ### Patch Changes