+++ components/flo-legal-decision-tree/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @utrecht/flolegal-decision-tree-css
 
+## 2.1.2
+
+### Patch Changes
+
+- e09112a: Update de Flo Legal Decision Tree client en CSS naar versie 1.17.0 om bugfixes en verbeteringen door te voeren.
+
+## 2.1.1
+
+### Patch Changes
+
+- 951bb2c: Align the flo-legal style with Figma
+
 ## 2.1.0
 
 ### Minor Changes