+++ packages/component-library-vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @utrecht/component-library-vue
 
+## 3.0.0
+
+### Major Changes
+
+- edac79f: - badge-status is no longer styled through common badge tokens but has its own set of tokens
+  - badge-status has been renamed to status-badge, use @utrecht/status-badge-css instead of @utrecht/badge-status-css
+
 ## 2.4.1
 
 ### Patch Changes