+++ packages/storybook-helpers/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @utrecht/storybook-helpers
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [341a0b7]
+  - @utrecht/design-tokens@5.0.1
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [edac79f]
+- Updated dependencies [0fcdfbf]
+  - @utrecht/design-tokens@5.0.0
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [f02b554]
+- Updated dependencies [55988ac]
+  - @utrecht/design-tokens@4.2.0
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [e7b59b3]
+- Updated dependencies [951bb2c]
+- Updated dependencies [fd2db7b]
+  - @utrecht/design-tokens@4.1.1
+
 ## 0.0.3
 
 ### Patch Changes